Rename `quadrature._base.py` to `_scheme.py`
