+++ src/felupe/quadrature/__init__.py
@@ -1,4 +1,4 @@
-from ._base import Scheme
+from ._scheme import Scheme
 from ._gausslegendre import GaussLegendre, GaussLegendreBoundary
 from ._tetra import Tetrahedron
 from ._triangle import Triangle
